Added conditionals in download method to not re-download existing files

+++ src/opendr/perception/object_detection_2d/yolov3/yolov3_learner.py
@@ -572,21 +572,38 @@ class YOLOv3DetectorLearner(Learner):
                                     "yolo_voc.json")
             if verbose:
                 print("Downloading metadata...")
-            urlretrieve(file_url, os.path.join(path, "yolo_default.json"))
+            if not os.path.exists(os.path.join(path, "yolo_default.json")):
+                urlretrieve(file_url, os.path.join(path, "yolo_default.json"))
+                if verbose:
+                    print("Downloaded metadata json.")
+            else:
+                if verbose:
+                    print("Metadata json file already exists.")
 
             if verbose:
                 print("Downloading params...")
             file_url = os.path.join(url, "pretrained", "yolo_voc",
                                          "yolo_voc.params")
 
-            urlretrieve(file_url,
-                        os.path.join(path, "yolo_voc.params"))
+            if not os.path.exists(os.path.join(path, "yolo_voc.params")):
+                urlretrieve(file_url, os.path.join(path, "yolo_voc.params"))
+                if verbose:
+                    print("Downloaded params.")
+            else:
+                if verbose:
+                    print("Params file already exists.")
 
         elif mode == "images":
             file_url = os.path.join(url, "images", "cat.jpg")
             if verbose:
                 print("Downloading example image...")
-            urlretrieve(file_url, os.path.join(path, "cat.jpg"))
+            if not os.path.exists(os.path.join(path, "cat.jpg")):
+                urlretrieve(file_url, os.path.join(path, "cat.jpg"))
+                if verbose:
+                    print("Downloaded example image.")
+            else:
+                if verbose:
+                    print("Example image already exists.")
 
         elif mode == "test_data":
             os.makedirs(os.path.join(path, "test_data"), exist_ok=True)
@@ -596,17 +613,35 @@ class YOLOv3DetectorLearner(Learner):
             file_url = os.path.join(url, "test_data", "train.txt")
             if verbose:
                 print("Downloading filelist...")
-            urlretrieve(file_url, os.path.join(path, "test_data", "train.txt"))
+            if not os.path.exists(os.path.join(path, "test_data", "train.txt")):
+                urlretrieve(file_url, os.path.join(path, "test_data", "train.txt"))
+                if verbose:
+                    print("Downloaded filelist.")
+            else:
+                if verbose:
+                    print("Filelist already exists.")
             # download image
             file_url = os.path.join(url, "test_data", "Images", "000040.jpg")
             if verbose:
                 print("Downloading image...")
-            urlretrieve(file_url, os.path.join(path, "test_data", "Images", "000040.jpg"))
+            if not os.path.exists(os.path.join(path, "test_data", "Images", "000040.jpg")):
+                urlretrieve(file_url, os.path.join(path, "test_data", "Images", "000040.jpg"))
+                if verbose:
+                    print("Downloaded image.")
+            else:
+                if verbose:
+                    print("Image already exists.")
             # download annotations
             file_url = os.path.join(url, "test_data", "Annotations", "000040.jpg.txt")
             if verbose:
                 print("Downloading annotations...")
-            urlretrieve(file_url, os.path.join(path, "test_data", "Annotations", "000040.jpg.txt"))
+            if not os.path.exists(os.path.join(path, "test_data", "Annotations", "000040.jpg.txt")):
+                urlretrieve(file_url, os.path.join(path, "test_data", "Annotations", "000040.jpg.txt"))
+                if verbose:
+                    print("Downloaded annotations.")
+            else:
+                if verbose:
+                    print("Annotations already exist.")
 
     def optimize(self, target_device):
         """This method is not used in this implementation."""
